Rename command_creation to command_creation_config; move PIPER_CL_TEMPLATE to top of file
